+++ src/client/cluster/mesher/index.ts
@@ -1,0 +1,1 @@
+export { ChunkMesher } from "./ChunkMesher";