+++ src/client/game/Game.ts
@@ -4,36 +4,34 @@ import * as Babylon from "babylonjs";
 import { Vector3 } from "babylonjs";
 
 import { GUIManager, IGUIManager } from "@client/gui";
-import { ClusterClient, IClusterClient } from "@client/cluster-client";
-import { IGame } from ".";
+import { ClusterClient, IClusterClient } from "@client/cluster";
 import { IView } from "@client/view";
 
 import { AssetManager, IAssetManager } from "@client/assets";
 import { PlayerMotor, IPlayerMotor } from "@client/movement";
 
-// TODO: REMOVE THIS IMPORT
-import { ContinuousClusterGenerator, IClusterGenerator } from "@server/cluster-gen";
+import { IClusterData } from "@share/data/cluster-data";
+import { ContinuousClusterGenerator } from "@server/cluster-gen";
 
 /**
  * The runner class for all game logic.
  */
-export class Game implements IGame {
+export class Game {
 
   // Overhead
-  private engine: Babylon.Engine;
-  private scene: Babylon.Scene;
+  private _engine: Babylon.Engine;
+  private _scene: Babylon.Scene;
 
   // Lighting
-  private sun: Babylon.DirectionalLight;
-  private hemisphericLight: Babylon.HemisphericLight;
-  private shadowGenerator: Babylon.ShadowGenerator | null = null;
+  private _sun: Babylon.DirectionalLight;
+  private _hemisphericLight: Babylon.HemisphericLight;
+  private _shadowGenerator: Babylon.ShadowGenerator | null = null;
 
   // Game elements
-  private view: IView;
-  private cluster: IClusterClient;
-  private motor: IPlayerMotor;
-  private gui: IGUIManager;
-  private assetManager: IAssetManager | null = null;
+  private _cluster: IClusterClient;
+  private _motor: IPlayerMotor;
+  private _gui: IGUIManager;
+  private _assetManager: IAssetManager | null = null;
 
   /**
    * Create a new Game.
@@ -41,51 +39,54 @@ export class Game implements IGame {
    * @param debugMode Show Babylon debug GUI.
    */
   constructor(
-    view: IView,
+    private _view: IView,
+    //clusterData: IClusterData,
     debugMode = false
   )
   {
-    this.view = view;
-    this.engine = new Babylon.Engine(view.getCanvas(), true);
+    this._engine = new Babylon.Engine(_view.getCanvas(), debugMode);
 
     // Set up the scene
-    this.scene = this._initScene(debugMode);
+    this._scene = this._initScene(debugMode);
     this._addEventListeners();
 
     // Set up lighting
-    this.sun = new Babylon.DirectionalLight("sun", new Vector3(-1, -1, -1), this.scene);
-    this.sun.intensity = 1.2;
-    this.sun.position = new Vector3(100, 100, 100);
-    this.hemisphericLight = new Babylon.HemisphericLight("ambience", new Vector3(-1, 1, -1), this.scene);
-    this.hemisphericLight.intensity = 0.3;
+    this._sun = new Babylon.DirectionalLight("sun", new Vector3(-1, -1, -1), this._scene);
+    this._sun.intensity = 1.2;
+    this._sun.position = new Vector3(100, 100, 100);
+    this._hemisphericLight = new Babylon.HemisphericLight(
+      "ambience", new Vector3(-1, 1, -1), this._scene);
+    this._hemisphericLight.intensity = 0.3;
 
     // Init asset manager
-    this.assetManager = new AssetManager(this.scene);
+    this._assetManager = new AssetManager(this._scene);
 
     // Init shadow generator
     //this.shadowGenerator = new Babylon.ShadowGenerator(1024, this.sun);
     //this.shadowGenerator.usePoissonSampling = true;
 
     // Create world cluster
-    const clusterGenerator: IClusterGenerator = new ContinuousClusterGenerator();
-    const clusterData = clusterGenerator.createWorldCluster();
+    const clusterData = ContinuousClusterGenerator.createWorldCluster(); //TODO: REMOVE THIS LINE
+
+    // Get world data from the server
+
     //this.cluster = new ClusterClient(clusterData, this.shadowGenerator, this.assetManager);
-    this.cluster = new ClusterClient(clusterData, null, this.assetManager);
-    this.cluster.remesh();
+    this._cluster = new ClusterClient(clusterData, null, this._assetManager);
+    this._cluster.remesh();
 
     // Create local player motor
-    this.motor = new PlayerMotor(
-      view.getCanvas(), this.engine, this.scene, this.cluster, new Vector3(10, 22, 10), true);
+    this._motor = new PlayerMotor(
+      _view.getCanvas(), this._engine, this._scene, this._cluster, new Vector3(50, 70, -50), true);
 
-    this.gui = new GUIManager();
-    this.gui.mainMenuGui();
+    this._gui = new GUIManager();
+    this._gui.mainMenuGui();
     //this.gui.gameGui();
 
     // Run engine render loop
-    const fpsElement = view.getFpsElement();
-    const scene = this.scene;
-    const engine = this.engine;
-    this.engine.runRenderLoop(() => {
+    const fpsElement = _view.getFpsElement();
+    const scene = this._scene;
+    const engine = this._engine;
+    this._engine.runRenderLoop(() => {
       scene.render();
       fpsElement.innerHTML = engine.getFps().toFixed();
     });
@@ -97,7 +98,7 @@ export class Game implements IGame {
   _initScene(debugMode: boolean = false): Babylon.Scene {
 
     // Set up scene
-    const scene = new Babylon.Scene(this.engine);
+    const scene = new Babylon.Scene(this._engine);
     if (debugMode) {
       scene.debugLayer.show();
     }
@@ -111,15 +112,15 @@ export class Game implements IGame {
    * Add event listeners.
    */ 
   _addEventListeners() {
-    const engine = this.engine;
+    const engine = this._engine;
 
     // Window resize listener
     window.addEventListener("resize", () => {
       engine.resize();
     });
 
-    this.scene.onPointerDown = event => {
-      this.engine.enterPointerlock();
+    this._scene.onPointerDown = event => {
+      this._engine.enterPointerlock();
     };
   }
 }