+++ src/server/cluster-gen/ContinuousClusterGenerator.ts
@@ -1,52 +1,38 @@
-import { Block } from "@share/utility";
+import { ContinuousBiomeGenerator, IContinuousBiomeGenerator, OneBiomeGenerator } from "./biome-gen";
+import { BiomeDependentHillGenerator, InterpolatedHillGenerator, IHillGenerator } from "./hill-gen";
 import { ClusterData, IClusterData } from "@share/data/cluster-data";
-import { IClusterGenerator } from "./IClusterGenerator";
-import { Biome } from "@share/utility/Biome";
-import { ContinuousBiomeGenerator, IContinuousBiomeGenerator } from "./biome-gen/continuous-biome-gen";
-import { BiomeDependentHillGenerator } from "./hill-gen/biome-dependent/BiomeDependentHillGenerator";
 import { a } from "@share/data/coordinate/CoordinateGenerators";
-import { InterpolatedHillGenerator } from "./hill-gen/InterpolatedHillGenerator";
+import { Block, Biome, getBlockFromBiome } from "@share/utility";
 
-const WORLD_WIDTH = 200;
+const WORLD_WIDTH = 100;
 const BIOME_WIDTH = 50;
-const BIOME_BORDER_GRADE = 0.5;
 
 /**
  * Generator for cluster data.
  */
-export class ContinuousClusterGenerator implements IClusterGenerator {
+export class ContinuousClusterGenerator {
 
   /**
    * Generate a standard world.
    */
-  createWorldCluster(): IClusterData {
+  static createWorldCluster(): IClusterData {
     const cluster = new ClusterData();
 
     // Generate biomes
-    const biomeGen: IContinuousBiomeGenerator = 
-      new ContinuousBiomeGenerator(WORLD_WIDTH, BIOME_WIDTH, BIOME_BORDER_GRADE);
+    //const biomeGen: IContinuousBiomeGenerator = new ContinuousBiomeGenerator(WORLD_WIDTH, BIOME_WIDTH);
+    const biomeGen: IContinuousBiomeGenerator = new OneBiomeGenerator(Biome.Grasslands);
 
     // Generate hills
-    const hillGen = new BiomeDependentHillGenerator(WORLD_WIDTH, biomeGen);
-    //const hillGen = new InterpolatedHillGenerator(biomeGen);
+    const hillGen: IHillGenerator = new BiomeDependentHillGenerator(WORLD_WIDTH, biomeGen);
+    //const hillGen: IHillGenerator = new InterpolatedHillGenerator(biomeGen);
 
     for (let x = 0; x < WORLD_WIDTH; x++) {
       for (let z = 0; z < WORLD_WIDTH; z++) {
 
         const y = hillGen.getYFromXZ(x, z);
+        const biome: Biome = biomeGen.getBiomesFromXZ(x, z).getTopBiome();
+        const topBlock = getBlockFromBiome(biome);
 
-        const biome: Biome = biomeGen.getTopBiomeFromXZ(x, z);
-        let topBlock: Block;
-        if (biome == Biome.Grasslands) {
-          topBlock = Block.Grass;
-        }
-        else if (biome == Biome.Desert) {
-          topBlock = Block.Sand;
-        }
-        else {
-          topBlock = Block.Stone;
-        }
-        
         // Set block column
         cluster.setBlock(a(x, y, z), topBlock);
         for (let i = y - 1; i >= 0; i--) {
@@ -59,7 +45,11 @@ export class ContinuousClusterGenerator implements IClusterGenerator {
       }
     }
 
-    /*for (let point of hillGen.coords) {
+    /*for (let point of biomeGen.points) {
+      cluster.setBlock(a(point[0], 100, point[1]), BiomeBlocks.getBlockFromBiome(point[2]));
+    }
+
+    for (let point of hillGen.coords) {
       cluster.setBlock(point, Block.Stone);
     }*/