+++ src/server/socket/socket-incoming/ISocketIncoming.ts
@@ -1,8 +1,0 @@
-
-/**
- * Handle incoming messages from a client and
- * send them to the IGameServer.
- */
-export interface ISocketIncoming {
-
-}