+++ src/server/socket/socket-incoming/index.ts
@@ -1,2 +1,0 @@
-export { SocketIncoming } from "./SocketIncoming";
-export { ISocketIncoming } from "./ISocketIncoming";