+++ changelog.md
@@ -1,3 +1,16 @@
+## Backlog
+* Send world data between client and server
+* Sound FX
+* Fix player movement behavior (going under arches, placing blocks under self, etc.)
+* Make asset manager static
+* Remove coordinate system and greatly simplify grid system
+
+## 0.0.4
+
+* Fixed lines between blocks
+* Improved top-down biome generation (inverse instead of logistic)
+
+
 ## 0.0.3
 
 * Add complex terrain generation with logistic curves